Modifying HNM code before integration with TTS: - Updated speech frame representation - Removed non-used class members - Moved LP parameters into noise part representation - Converted double parameters to float and tested
git-svn-id: https://mary.opendfki.de/repos/trunk@1934 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/marytts/util/math/ComplexArray.java
@@ -72,7 +72,7 @@ public class ComplexArray
     
     public ComplexNumber get(int index)
     {
-        return new ComplexNumber(real[index], imag[index]);
+        return new ComplexNumber((float)real[index], (float)imag[index]);
     }
 }
